efi: Move efi module to its own directory
And move the EFI allocator into it. As there will be more EFI specific
functionality coming it makes sense to group that together.
